Rearrange the directories
