+++ includes/dowithdraw.php
@@ -1,14 +1,14 @@
 <?PHP
-include_once('config.php');
+include_once('../config.php');
 $user = $_POST['user'];
 $amount = $_POST['amount'];
 $receiver = $_POST['receiver'];
  $balance = 0;
  $balancedb = 0;
   $sql = "SELECT address, currency FROM address WHERE user='$user'";
-  
+
                                 $result = $con->query($sql);
-                               
+
                                 while ($row = $result->fetch_assoc()){
                                     $address = $row['address'];
                                     $currency = $row['currency'];
@@ -16,16 +16,16 @@ $receiver = $_POST['receiver'];
                                     $data = json_decode($url, TRUE);
                                     $confirmed = $data['data']['confirmed_balance'];
                                     $balance = $balance + $confirmed;
-                                  
+
                                 }
-                               
+
                                  $sql2 = "SELECT balance, address FROM balance WHERE user='$user'";
                                 $result2 = $con->query($sql2);
                                 if ($result2->num_rows > 0){
                                 while($row2 = $result2->fetch_assoc()){
                                     $bconfirmed = $row2['balance'];
                                     $balancedb = $balancedb + $bconfirmed;
-                                    
+
                                 }
 }
 $balancedb = $balancedb - 0.0005;
@@ -42,7 +42,7 @@ if ($amount == $balance or $amount < $balance){
 	$tx = $result['result']['send_txid'];
 	    $sql3 = "INSERT INTO history (user, amount, id, tx, receiver) VALUES ('$user', '$amount', '$id', '$tx', '$receiver')";
 	    if ($con->query($sql3) === TRUE){
-	        
+
 	    }
 		print 'Withdrawal created with ID: '.$result['result']['id'];
 	} else {
@@ -50,21 +50,21 @@ if ($amount == $balance or $amount < $balance){
 	}
 } else{
 	echo "Sorry, but the amount you've entered is more than your balance.";
-    } 
+    }
     }else{
 	echo "Transaction NOT completed. Our system is trying to verify your transaction. Please try again in few minutes.";
 }
              <?PHP
-include_once('config.php');
+include_once('../config.php');
 $user = $_POST['user'];
 $amount = $_POST['amount'];
 $receiver = $_POST['receiver'];
  $balance = 0;
  $balancedb = 0;
   $sql = "SELECT address, currency FROM address WHERE user='$user'";
-  
+
                                 $result = $con->query($sql);
-                               
+
                                 while ($row = $result->fetch_assoc()){
                                     $address = $row['address'];
                                     $currency = $row['currency'];
@@ -72,16 +72,16 @@ $receiver = $_POST['receiver'];
                                     $data = json_decode($url, TRUE);
                                     $confirmed = $data['data']['confirmed_balance'];
                                     $balance = $balance + $confirmed;
-                                  
+
                                 }
-                               
+
                                  $sql2 = "SELECT balance, address FROM balance WHERE user='$user'";
                                 $result2 = $con->query($sql2);
                                 if ($result2->num_rows > 0){
                                 while($row2 = $result2->fetch_assoc()){
                                     $bconfirmed = $row2['balance'];
                                     $balancedb = $balancedb + $bconfirmed;
-                                    
+
                                 }
 }
 $balancedb = $balancedb - 0.0005;
@@ -98,7 +98,7 @@ if ($amount == $balance or $amount < $balance){
 	$tx = $result['result']['send_txid'];
 	    $sql3 = "INSERT INTO history (user, amount, id, tx, receiver) VALUES ('$user', '$amount', '$id', '$tx', '$receiver')";
 	    if ($con->query($sql3) === TRUE){
-	        
+
 	    }
 		print 'Withdrawal created with ID: '.$result['result']['id'];
 	} else {
@@ -106,8 +106,8 @@ if ($amount == $balance or $amount < $balance){
 	}
 } else{
 	echo "Sorry, but the amount you've entered is more than your balance.";
-    } 
+    }
     }else{
 	echo "Transaction NOT completed. Our system is trying to verify your transaction. Please try again in few minutes.";
 }
-             
+
