+++ includes/withdraw.php
@@ -1,16 +1,20 @@
-<?PHP 
-include_once('config.php');
+<?PHP
+include_once('../config.php');
 session_start();
 $filename = basename($_SERVER['REQUEST_URI']);
+
 if ($_SESSION['login']==1){
-    
+
 ?>
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml">
 <head>
+    <?PHP
+    include_once('menu.php');
+    ?>
       <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Address | Simple BTC Wallet</title>
+    <title>Withdrawal | Simple BTC Wallet</title>
 	<!-- BOOTSTRAP STYLES-->
     <link href="assets/css/bootstrap.css" rel="stylesheet" />
      <!-- FONTAWESOME STYLES-->
@@ -22,7 +26,7 @@ if ($_SESSION['login']==1){
    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 </head>
 <body>
-     
+
            <style>
 .alert {
     padding: 20px;
@@ -45,7 +49,7 @@ if ($_SESSION['login']==1){
     color: black;
 }
 </style>
-         
+
     <div id="wrapper">
          <div class="navbar navbar-inverse navbar-fixed-top">
             <div class="adjust-nav">
@@ -57,9 +61,9 @@ if ($_SESSION['login']==1){
                     </button>
                      <font color="white"><h1>Simple BTC Wallet</h1></font>
                 </div>
-              
+
                  <span class="logout-spn" >
-                  <a id="logout" class="btn btn-danger btn-lg btn-block" style="color:#fff;">LOGOUT</a>  
+                  <a id="logout" class="btn btn-danger btn-lg btn-block" style="color:#fff;">LOGOUT</a>
 
                 </span>
             </div>
@@ -68,18 +72,16 @@ if ($_SESSION['login']==1){
         <nav class="navbar-default navbar-side" role="navigation">
             <div class="sidebar-collapse">
                 <ul class="nav" id="main-menu">
-                 
 
-
-                    <li>
+ <li>
                         <a href="dash.php" ><i class="fa fa-desktop "></i>Dashboard </a>
                     </li>
-                   
 
-                    <li class="active-link">
+
+                    <li>
                         <a href="address.php"><i class="fa fa-book "></i>Addresses</a>
                     </li>
-                    <li>
+                    <li class="active-link">
                         <a href="withdraw.php"><i class="fa fa-paper-plane "></i>Withdraw</a>
                     </li>
                     <li>
@@ -97,59 +99,67 @@ if ($_SESSION['login']==1){
             <div id="page-inner">
                 <div class="row">
                     <div class="col-md-12">
-                     <h2>Addresses</h2>   
+                     <h2>Withdraw</h2>
                     </div>
-                </div>              
+                </div>
                  <!-- /. ROW  -->
                   <hr />
-                   
-  
- <div id="alert">
-     
- </div>
 
-                   <table class="table table-striped table-bordered table-hover">
-                            <thead>
-                                <tr>
-                                    <th>Address</th>
-                                    <th>Currency</th>
-                                    
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <?PHP
-                                $user = $_SESSION['user'];
-                               
-                                $sql = "SELECT address, currency FROM address WHERE user='$user'";
+
+ <div id="alert">
+
+ </div>
+ <?PHP
+ $user = $_SESSION['user'];
+ $balance = 0;
+ $balancedb = 0;
+  $sql = "SELECT address, currency FROM address WHERE user='$user'";
                                 $result = $con->query($sql);
                                 while ($row = $result->fetch_assoc()){
                                     $address = $row['address'];
                                     $currency = $row['currency'];
-                                    echo "<tr> <td>$address</td><td>$currency</td></tr>";
+                                    $url = file_get_contents("https://chain.so/api/v2/get_address_balance/BTC/$address");
+                                    $data = json_decode($url, TRUE);
+                                    $confirmed = $data['data']['confirmed_balance'];
+                                    $balance = $balance + $confirmed;
                                 }
-                                ?>
-                                
-                            </tbody>
-                        </table>
+                                 $sql2 = "SELECT balance, address FROM balance WHERE user='$user'";
+                                $result2 = $con->query($sql2);
+
+                                while($row2 = $result2->fetch_assoc()){
+                                    $bconfirmed = $row2['balance'];
+                                    $balancedb = $balancedb + $bconfirmed;
+                                }
+
+                                echo "Currenct Confirmed Balance: $balance BTC";
+
+ ?>
+<p>Fee: 0.0005 BTC</p>
+                  <input class="form-control"  name ="amount" placeholder='Amount' >
+                  <br>
+                  <input class="form-control"  name ="currency" placeholder='Currency' >
+                  <br>
+                  <input class="form-control" name="to" placeholder="Receiver">
+
                          <input class="form-control"  name ="user" type="hidden"  value='<?PHP echo $_SESSION["user"]; ?>'/>
-                        
-                        <a class="btn btn-danger btn-lg btn-block" id="new">New address</a>
-                 <!-- /. ROW  -->           
+                         <br>
+                        <a class="btn btn-danger btn-lg btn-block" id="sub">Withdraw</a>
+                 <!-- /. ROW  -->
     </div>
              <!-- /. PAGE INNER  -->
             </div>
          <!-- /. PAGE WRAPPER  -->
         </div>
     <div class="footer">
-      
-    
-           <div class="row">
+
+
+             <div class="row">
                 <div class="col-lg-12" >
                     &copy;  2018 yourdomain.com | Theme by: <a href="http://binarytheme.com" style="color:#fff;" target="_blank">www.binarytheme.com</a> | Script by: <a href="http://megacrypto.online" style="color:#fff;" target="_blank">Megacrypto</a>
                 </div>
             </div>
         </div>
-          
+
 
      <!-- /. WRAPPER  -->
     <!-- SCRIPTS -AT THE BOTOM TO REDUCE THE LOAD TIME-->
@@ -159,40 +169,46 @@ if ($_SESSION['login']==1){
     <script src="assets/js/bootstrap.min.js"></script>
       <!-- CUSTOM SCRIPTS -->
     <script src="assets/js/custom.js"></script>
-    
+
    <script>
-       	$('#new').click(function() {
-			      
+       	$('#sub').click(function() {
+
 			    var user = $('input[name=user]').val();
+			    var amount = $('input[name=amount]').val();
+			    var receiver = $('input[name=to]').val();
     $.ajax({
-        url: 'new.php',
+
+        url: 'dowithdraw.php',
         type: 'POST',
         data: {
-           user:user
-           
+           user:user,
+           amount:amount,
+           receiver:receiver
+
         },
-        success: function(msg) {
-             $("#alert").append('  <div class="alert"> <span class="closebtn" onclick="this.parentElement.style.display="none";">&times;</span> <b>A new address has been successfully generated!</b> </div>');
-           
-          
-        }               
+        success: function(data) {
+             alert(data);
+
+
+        }
     });
+
 });
 $('#logout').click(function() {
-			      
-			    
+
+
     $.ajax({
         url: 'logout.php',
         type: 'POST',
         data: {
            user:'h'
-           
+
         },
         success: function(logout) {
-     $("#alert").html(logout);
-           
-          
-        }               
+    $("#alert").html(logout);
+
+
+        }
     });
 });
    </script>
